TsqeuryFactory moved to String component folder
- TsqueryFactory works with strings and seperating this factory to seperate folder is unnecessary
